Initial commit. Basic room shading, visualisation and artnet working.
This sets up the general skeleton of the project, ready for super basic
2D visualisation and DMX output via artnet over sACN.

The 3D shading API is also setup ready to go, currently with a basic
test pattern.

Fixtures are currently hard coded, but eventually we'll want to add a UI
for custom DMX address assignment.

Existing LED strip visualisation will probably change and the associated
DMX is not yet being sent.