Update wash positions to match blueprint. Add topdown window view.
A new `layout` module has been added which contains some functions and
consts describing the layout of the structure and lighting in metres.
The module also contains a function for converting from topdown
coordinates in metres to shader coordinates that are normalised over the
width of the venue and whose origin is at the centre of the floor under
the LED strips.

The old vis window still needs to be removed, but it has been handy for
ensuring the shader coords are all still working well.

Follow-up TODO:

- Add rough "height" function for wash targets based on whether they hit
  the floor or wall.
- LED strip visualisation.