Add rough wash heights for wash lights that hit the walls
This should allow for at least a bit of movement over the y axis.

Also added a lil feature so that when holding down left shift and
hovering the mouse over the topdown window, the x and z shader
coordinates will be displayed. Useful for double checking how the
shaders will affect things.