Add a `last_color` field to the Vertex type
This new field provides the last colour produced by the shader for this
vertex.

@JoshuaBatty there's a `let lerp_amt = 1.0;` line towards the end of the
shader that we can replace with whatever uniform value you want to use
to control the wash light interpolation.

Note the current behaviour isn't quite correct as the blending is
occurring outside of the shader. If we can move the blending inside the
shader (e.g. moving the blend mode into the uniforms struct and matching
inside the shader instead) then that should fix this issue.