Separate application of fade-to-black from shader outputs
Currently the wash interpolation works by interpolating from the
previous colour stored in its colour buffer. This meant that washes
would always be overly dark when a long interpolation and fade value
less than one was used. This fixes the issue by storing the output of
the fade-to-black separately from the original shader output.