Make shader params GUI handling more generic
This adds a `gui::Params` trait implemented for each of the shader
parameters types that allows for instantiating the shader and colorising
GUI based on the number of parameters and their types.

Closes #17

Still have to solve the issue where both left and right use the same set
of parameters when the same shader is selected. Should be pretty easy to
just make `left` and `right` instances of the `ShaderParams` to be
stored in the config though. Will do that next.