Commit config with presets. Add shader and preset with button demo.
This commit includes the config.json into the gitignore so that we can
start committing presets. I've added 6 or 7 presets to kick things off,
but feel free to rename/remove them or whatevs if you have a flow in
mind for the night!

This also re-adds in the original wash test shader I added and uses it
to demonstrate the affect of button presses. You can trigger the radial
glow burst effect by pressing the cycle button on the korg (or spacebar
which pretends to be the cycle button). This is just to give an idea of
how it can potentially spice up the shaders a bit and add a bit of
musical/spatial interactivity :+1: